+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fixed
 - read_only was forced False on nested form elements when unset on the parent
 - read_only working for ChipGroup and MultiSelect
+- form_cols is now passed to the form fields which fixes issues with modal renders
 
 ## [0.10.0] - 2024-12-04
 ### Added